refactor(core): Placeholder image to FakeImg class
* placehold.mochizuki.moe is obsoleted.
